refactor: move hooks from store to hooks folder
