+++ apps/web/src/components/auth/useAcceptInvite.ts
@@ -6,7 +6,7 @@ import * as Sentry from '@sentry/react';
 
 import { api } from '../../api/api.client';
 import { useAuthContext } from '../../store/authContext';
-import { applyToken } from '../../store/useAuthController';
+import { applyToken } from '../../hooks/useAuthController';
 
 export function useAcceptInvite() {
   const { setToken } = useAuthContext();
