+++ apps/web/src/components/templates/TemplateSMSEditor.tsx
@@ -3,7 +3,7 @@ import { Control, Controller, useFormContext } from 'react-hook-form';
 import { LackIntegrationError } from './LackIntegrationError';
 import type { IForm } from './formTypes';
 import { Textarea } from '../../design-system';
-import { useEnvController } from '../../store/useEnvController';
+import { useEnvController } from '../../hooks/useEnvController';
 import { VariableManager } from './VariableManager';
 import { useVariablesManager } from '../../hooks/useVariablesManager';
