+++ apps/web/src/components/templates/email-editor/EmailCustomCodeEditor.tsx
@@ -5,7 +5,7 @@ import { addCompleter } from 'ace-builds/src-noconflict/ext-language_tools';
 import { Card } from '@mantine/core';
 import { SystemVariablesWithTypes, HandlebarHelpers } from '@novu/shared';
 import { colors } from '../../../design-system';
-import { useEnvController } from '../../../store/useEnvController';
+import { useEnvController } from '../../../hooks/useEnvController';
 
 export function EmailCustomCodeEditor({
   onChange,
