+++ apps/web/src/design-system/template-button/ChannelButton.tsx
@@ -10,7 +10,7 @@ import { useStyles } from './TemplateButton.styles';
 import { colors, shadows } from '../config';
 import { DotsHorizontal, Edit, Trash } from '../icons';
 import { When } from '../../components/utils/When';
-import { useEnvController } from '../../store/useEnvController';
+import { useEnvController } from '../../hooks/useEnvController';
 import { getChannel, NodeTypeEnum } from '../../pages/templates/shared/channels';
 
 import { useViewport } from 'react-flow-renderer';
