+++ apps/web/src/hooks/useLayouts.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { useEnvController } from '../store/useEnvController';
+import { useEnvController } from '../hooks/useEnvController';
 import { QueryKeys } from '../api/query.keys';
 import { getLayoutsList } from '../api/layouts';
 import { ILayoutEntity } from '@novu/shared';
