+++ apps/web/src/hooks/useTemplates.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { INotificationTemplate } from '@novu/shared';
 
-import { useEnvController } from '../store/useEnvController';
+import { useEnvController } from '../hooks/useEnvController';
 import { getNotificationsList } from '../api/notification-templates';
 
 export function useTemplates(page = 0, limit = 10) {
