+++ apps/web/src/pages/brand/tabs/LayoutsListPage.tsx
@@ -13,7 +13,7 @@ import { useLayouts } from '../../../hooks/useLayouts';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { deleteLayoutById } from '../../../api/layouts';
 import { errorMessage, successMessage } from '../../../utils/notifications';
-import { useEnvController } from '../../../store/useEnvController';
+import { useEnvController } from '../../../hooks/useEnvController';
 import { QueryKeys } from '../../../api/query.keys';
 
 const enum ActivePageEnum {
