+++ apps/web/src/pages/settings/tabs/EmailSettings.tsx
@@ -7,7 +7,7 @@ import { colors, Text, Input, Tooltip, Button } from '../../../design-system';
 import { Check, CheckCircle, Copy } from '../../../design-system/icons';
 import React, { useEffect } from 'react';
 import { Control, Controller, useForm } from 'react-hook-form';
-import { useEnvController } from '../../../store/useEnvController';
+import { useEnvController } from '../../../hooks/useEnvController';
 import { useMutation } from '@tanstack/react-query';
 import { updateDnsSettings } from '../../../api/environment';
 import { showNotification } from '@mantine/notifications';
