+++ apps/web/src/pages/settings/tabs/components/ConfirmRegenerationModal.tsx
@@ -1,6 +1,6 @@
 import { Group, Modal, useMantineTheme } from '@mantine/core';
 import { Button, colors, shadows, Title, Text } from '../../../../design-system';
-import { useEnvController } from '../../../../store/useEnvController';
+import { useEnvController } from '../../../../hooks/useEnvController';
 
 export function ConfirmRegenerationModal({
   isOpen,
