+++ apps/web/src/pages/templates/workflow/DelayMetadata.tsx
@@ -4,7 +4,7 @@ import { Controller, useFormContext } from 'react-hook-form';
 
 import { Input, Select } from '../../../design-system';
 import { inputStyles } from '../../../design-system/config/inputs.styles';
-import { useEnvController } from '../../../store/useEnvController';
+import { useEnvController } from '../../../hooks/useEnvController';
 import { When } from '../../../components/utils/When';
 
 export const DelayMetadata = ({ control, index }) => {
