+++ apps/web/src/pages/templates/workflow/StepActiveSwitch.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Controller } from 'react-hook-form';
 import { Switch } from '../../../design-system';
-import { useEnvController } from '../../../store/useEnvController';
+import { useEnvController } from '../../../hooks/useEnvController';
 
 export const StepActiveSwitch = ({ control, index }) => {
   const { readonly } = useEnvController();
