+++ apps/web/src/store/authContext.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { IOrganizationEntity, IUserEntity, IJwtPayload } from '@novu/shared';
-import { useAuthController } from './useAuthController';
+import { useAuthController } from '../hooks/useAuthController';
 
 type UserContext = {
   token: string | null;
